fix: stabilize zfs builds (#29)
TL;DR - zfs builds are more reliable because they now only use the CoreOS image version upon which they mean to install instead of attempting to guess a matching "Fedora" version for a builder.


I investigated continued instability of ZFS builds thinking it could be related to the F38 release. It may have been in a manner of speaking. 
I discovered the method of using a Fedora image as a builder was more complex than needed and was the actual culprit.

Switched to using CoreOS image itself as the builder, which primarily required using rpm-ostree (instead of dnf) and the hack to allow alternatives to function (mkdir -p /var/lib/alternatives).