fix: change docker-compose installation
On Fedora 38 (testing) the `docker-compose` RPM is no longer
installable. It depends on `python3-dockerpty` which is not maintained
and may not be added to the Fedora 38 repos. I've replaced the old
python version which was installing from repos with a static binary
sourced from docker's official container image for the tool.