fix: enable testing repos on testing stream (#35)
I had assumed the testing repos were enabled by default on the FCOS
testing stream, but it seems they are not. This change enables them, but
only after ensuring they are not already enabled. This should correct
build failures which occur on testing when packages are being worked on
and in testing repos but not yet moved to stable.