feat: install docker-compose as cli plugin, add buildx
Some users had noticed that `docker-compose` was not usable as a plugin
(eg, `docker compose`). This fixes that by installing it the way moby
does upstream, plus adds `docker-buildx`. Both are versioned now rather
than being `latest`, and version matched to the moby release.

As of this writing, FCOS next is still on moby 24.0.5.
Some details on versions:
- https://fedoraproject.org/coreos/release-notes?arch=x86_64&stream=testing
- https://github.com/docker/buildx/releases
- https://github.com/docker/compose/releases

Closes: #141