fix: do not install pre-release packages
Two previous changes had added the fedora-coreos-pool.repo to mitigate
build issues with slightly older CoreOS RPM dependencies no longer
being available during builds of uCore. This was done without any
prioritization changes, and initially worked fine as Fedora 40 was the
newest release in the repo. Now it seems Fedora 41 packages are making
their way into the uCore images, which was NOT intended and not
desired. This removes that repo.

If it is ever needed again, will need to investigate deprioritizing so
normal repos are used except for missing packages.

Fixes: #174

Relates: #163
Relates: #167