fix: github release install silently fails if no urls found
This copies the fix from ublue-os/main#571 to ensure that the github
release install script fails properly when nothing matches the filter.

The bug fixed by this also relates to ublue-os/ucore#214 since that kind
of bug report will be caught by build failures in the future, thanks to
this fix.