feat: add tuned to ucore-minimal (and ucore and hci) images
This adds tuned with atomic profile but does not include the heavier
weak deps of kernel-tools and python3-perf. tuned.service is disabled
by default.

Though previous discussions did mention not installing this to the
minimal image, since the atomic profile specifically provides a
profile for VM guests, it seems reasonable there as well.